+++ src/frontend/src/features/participants/components/MuteEveryoneButton.tsx
