+++ src/frontend/src/features/participants/components/ParticipantsCollapsableList.tsx
