+++ src/frontend/src/features/participants/components/ParticipantsList.tsx
@@ -3,14 +3,14 @@ import { useParticipants } from '@livekit/components-react'
 
 import { Div, H } from '@/primitives'
 import { useTranslation } from 'react-i18next'
-import { ParticipantListItem } from '../../controls/Participants/ParticipantListItem'
-import { ParticipantsCollapsableList } from '../../controls/Participants/ParticipantsCollapsableList'
-import { HandRaisedListItem } from '../../controls/Participants/HandRaisedListItem'
-import { LowerAllHandsButton } from '../../controls/Participants/LowerAllHandsButton'
+import { ParticipantListItem } from './ParticipantListItem'
+import { ParticipantsCollapsableList } from './ParticipantsCollapsableList'
+import { HandRaisedListItem } from './HandRaisedListItem'
+import { LowerAllHandsButton } from './LowerAllHandsButton'
 import { WaitingParticipantListItem } from './WaitingParticipantListItem'
-import { useWaitingParticipants } from '@/features/rooms/hooks/useWaitingParticipants'
+import { useWaitingParticipants } from '@/features/participants/hooks/useWaitingParticipants'
 import type { Participant } from 'livekit-client'
-import { WaitingParticipant } from '@/features/rooms/api/listWaitingParticipants'
+import { WaitingParticipant } from '@/features/participants/api/listWaitingParticipants'
 import { MuteEveryoneButton } from './MuteEveryoneButton'
 
 // TODO: Optimize rendering performance, especially for longer participant lists, even though they are generally short.
