+++ src/frontend/src/features/participants/components/PinBadge.tsx
