+++ src/frontend/src/features/participants/components/menu/ParticipantMenuButton.tsx
