+++ src/frontend/src/features/participants/components/menu/items/PinMenuItem.tsx
