+++ src/frontend/src/features/participants/components/menu/items/RemoveMenuItem.tsx
